+++ src/main/java/lotto/exception/InvalidMoneyException.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.exception;
 
 public class InvalidMoneyException extends IllegalArgumentException {
 	public static final String NULL_OR_EMPTY = "null 또는 빈 문자열은 입력할 수 없습니다.";
