+++ src/main/java/lotto/exception/InvalidNumberException.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.exception;
 
 public class InvalidNumberException extends IllegalArgumentException {
 	public static final String OUT_OF_BOUND = "1이상 45이하의 정수를 입력해주세요.";
