+++ src/test/java/lotto/domain/NumberTest.java
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import lotto.exception.InvalidNumberException;
+
 public class NumberTest {
 	@DisplayName("Number 생성자에 1이상 45이하의 정수 입력이 들어올 때 객체 생성")
 	@ParameterizedTest
